Fix database locked errors in unit tests
